updated: fish configuration
added function to replace links with the actual file and make it
writable, this can be used to resolve the links generated by
home-manager and make debug edits to the file, since the file is going
to be replaced on the next rebuild anyways if it changed.

added starship and configuration for it.